add FC & types for navigation
