--- src/navigation/index.ts
@@ -1,3 +1,0 @@
-export {default as ApplicationNavigator} from './Application';
-export {default as MainNavigator, MainNavigatorStackName} from './Main';
-export * from './Root';
