+++ src/navigation/stacks/AuthStack.tsx
@@ -1,38 +1,55 @@
-import React from 'react';
+import React, {FC} from 'react';
 
-import {createStackNavigator} from '@react-navigation/stack';
+import {
+  createStackNavigator,
+  StackNavigationProp,
+} from '@react-navigation/stack';
 
-import {Login} from '~/screens/auth_stack/login_email';
-import {LoginWithPassword} from '~/screens/auth_stack/login_password';
-import {Register} from '~/screens/auth_stack/register';
-import {RegisterWithPass} from '~/screens/auth_stack/register_with_pass';
-import {Splash} from '~/screens/splash';
+import Login from '~/screens/AuthStack/Login';
+import LoginWithPassword from '~/screens/AuthStack/LoginPassword';
+import Register from '~/screens/AuthStack/Register';
+import RegisterWithPass from '~/screens/AuthStack/RegisterWithPass';
+
+export enum AuthStackName {
+  LOGIN = 'LOGIN',
+  LOGIN_WITH_PASS = 'LOGIN_WITH_PASS',
+  REGISTER = 'REGISTER',
+  REGISTER_WITH_PASS = 'REGISTER_WITH_PASS',
+}
+
+export type AuthStackParamList = {
+  [AuthStackName.LOGIN]: undefined;
+  [AuthStackName.LOGIN_WITH_PASS]: {
+    email: string;
+  };
+  [AuthStackName.REGISTER]: {
+    email: string;
+  };
+  [AuthStackName.REGISTER_WITH_PASS]: {
+    email: string;
+    code: string;
+  };
+};
+
+export type AuthStackNavigationProps = StackNavigationProp<AuthStackParamList>;
 
 const Stack = createStackNavigator();
-export const NameScreenAuthStack = {
-  SPLASH: 'SPLASH',
-  LOGIN: 'LOGIN',
-  LOGIN_WITH_PASS: 'LOGIN_WITH_PASS',
-  REGISTER: 'REGISTER',
-  INPUT_NAME: 'INPUT_NAME',
-  SELECT_CATEGORY: 'SELECT_CATEGORY',
-  REGISTER_WITH_PASS: 'REGISTER_WITH_PASS',
-};
 
-export const AuthStack = () => {
-  return (
-    <Stack.Navigator screenOptions={{headerShown: false}}>
-      <Stack.Screen name={NameScreenAuthStack.SPLASH} component={Splash} />
-      <Stack.Screen name={NameScreenAuthStack.LOGIN} component={Login} />
-      <Stack.Screen
-        name={NameScreenAuthStack.LOGIN_WITH_PASS}
-        component={LoginWithPassword}
-      />
-      <Stack.Screen name={NameScreenAuthStack.REGISTER} component={Register} />
-      <Stack.Screen
-        name={NameScreenAuthStack.REGISTER_WITH_PASS}
-        component={RegisterWithPass}
-      />
-    </Stack.Navigator>
-  );
-};
+const AuthStack: FC = () => (
+  <Stack.Navigator
+    initialRouteName={AuthStackName.LOGIN}
+    screenOptions={{headerShown: false}}>
+    <Stack.Screen name={AuthStackName.LOGIN} component={Login} />
+    <Stack.Screen
+      name={AuthStackName.LOGIN_WITH_PASS}
+      component={LoginWithPassword}
+    />
+    <Stack.Screen name={AuthStackName.REGISTER} component={Register} />
+    <Stack.Screen
+      name={AuthStackName.REGISTER_WITH_PASS}
+      component={RegisterWithPass}
+    />
+  </Stack.Navigator>
+);
+
+export default AuthStack;
