--- src/navigation/stacks/index.ts
@@ -1,1 +1,0 @@
-export {AuthStack, NameScreenAuthStack} from './AuthStack';
