+++ src/screens/MainStack/SelectCategory/Category/CategoryItem.tsx
