+++ src/screens/splash/index.tsx
@@ -1,12 +1,17 @@
-import React, {useEffect, useRef} from 'react';
+import React, {FC, useEffect, useRef} from 'react';
 import {Animated, Easing} from 'react-native';
 
+import {useNavigation} from '@react-navigation/native';
+import {useSelector} from 'react-redux';
 import styled from 'styled-components/native';
 
 import {SPLASH_IMAGE} from '~/assets';
 import BackgroundCommon from '~/components/BackgroundCommon';
-import {navigate} from '~/navigation';
-import {NameScreenAuthStack} from '~/navigation/stacks';
+import {
+  ApplicationStackName,
+  ApplicationStackNavigationProps,
+} from '~/navigation/Application';
+import {RootState} from '~/store';
 import {ContainerStyle} from '~/utils/Styles';
 
 const Container = styled.View`
@@ -17,12 +22,19 @@ const SplashImage = styled(Animated.Image)<{opacity: Animated.Value}>`
   opacity: ${({opacity}) => Number(opacity) || 0};
 `;
 
-export const Splash = () => {
+const Splash: FC = () => {
+  const {navigate} = useNavigation<ApplicationStackNavigationProps>();
+  const {isLoggedIn} = useSelector((state: RootState) => state.user.authData);
+
   const opacity = useRef(new Animated.Value(0)).current; // Initial value for opacity: 0
 
   useEffect(() => {
     const timeoutId: NodeJS.Timeout = setTimeout(() => {
-      navigate(NameScreenAuthStack.LOGIN, {});
+      if (isLoggedIn) {
+        navigate(ApplicationStackName.MAIN);
+      } else {
+        navigate(ApplicationStackName.AUTH);
+      }
     }, 3000);
 
     return () => {
@@ -49,3 +61,5 @@ export const Splash = () => {
     </BackgroundCommon>
   );
 };
+
+export default Splash;
