--- src/utils/navigation.ts
@@ -1,34 +1,0 @@
-import {TabActions} from '@react-navigation/native';
-import {StackActions} from '@react-navigation/routers';
-
-import {navigationRef} from '~/navigation/Root';
-
-export const navigation = () => navigationRef.current!;
-
-export const createNavigate =
-  <T extends object>(screenName: string) =>
-  (params?: T) => {
-    return navigation().navigate(screenName, params);
-  };
-
-export const createPush =
-  <T extends object>(screenName: string) =>
-  (params?: T) => {
-    return navigation().dispatch(StackActions.push(screenName, params));
-  };
-
-export const createReplace =
-  <T extends object>(screenName: string) =>
-  (params?: T) => {
-    return navigation().dispatch(StackActions.replace(screenName, params));
-  };
-
-export const goBack = () => navigation().goBack();
-
-export const popToTop = () => navigation().dispatch(StackActions.popToTop());
-
-export const jumpToTab = (tabName: string, params?: any) => {
-  const jumpToAction = TabActions.jumpTo(tabName, params);
-
-  return navigation().dispatch(jumpToAction);
-};
